+++ src/YCHR/DSL.hs
@@ -114,7 +114,7 @@ module YCHR.DSL
     Term,
     term,
     qterm,
-    quoted,
+    quote,
     var,
     atom,
     int,
@@ -161,7 +161,7 @@ import Control.Exception (throwIO)
 import Data.List.NonEmpty qualified as NE
 import Data.Map.Strict (Map)
 import Data.Text (Text)
-import YCHR.Convert (quoted)
+import YCHR.Convert (quote)
 import YCHR.Internal.Meta (metaHostCallRegistry)
 import YCHR.Internal.Parsed
 import YCHR.Internal.Runtime.Registry (HostCallRegistry, baseHostCallRegistry)