+++ test/YCHR/ConvertTest.hs
@@ -156,21 +156,21 @@ encodingTests =
         toTerm Dot @?= CompoundTerm (Unqualified "dot") [],
       testCase "Circle 3 (generic product)" $
         toTerm (Circle 3) @?= CompoundTerm (Unqualified "circle") [IntTerm 3],
-      testCase "quoted wraps a Term in term/1 unchanged" $
-        quoted (compound "plus" [int 2, int 3])
+      testCase "quote wraps a Term in quote/1 unchanged" $
+        quote (compound "plus" [int 2, int 3])
           @?= CompoundTerm
-            (Unqualified "term")
+            (Unqualified "quote")
             [CompoundTerm (Unqualified "plus") [IntTerm 2, IntTerm 3]],
-      testCase "quoted routes a non-Term through toTerm" $
-        quoted (Circle 3)
+      testCase "quote routes a non-Term through toTerm" $
+        quote (Circle 3)
           @?= CompoundTerm
-            (Unqualified "term")
+            (Unqualified "quote")
             [CompoundTerm (Unqualified "circle") [IntTerm 3]],
-      testCase "quoted nests" $
-        quoted (quoted (int 1))
+      testCase "quote nests" $
+        quote (quote (int 1))
           @?= CompoundTerm
-            (Unqualified "term")
-            [CompoundTerm (Unqualified "term") [IntTerm 1]]
+            (Unqualified "quote")
+            [CompoundTerm (Unqualified "quote") [IntTerm 1]]
     ]
 
 -- ---------------------------------------------------------------------------